Improve customer utilities and cleanup
