+++ tnp-backend/app/Http/Controllers/Api/V1/Customers/CustomerController.php
@@ -280,13 +280,8 @@ class CustomerController extends Controller
                 ->orderByDesc('cus_no')
                 ->first();
 
-            // Clean เบอร์โทรศัพท์ & Tax ID อัตโนมัติ
-            $fieldsToClean = ['cus_tel_1', 'cus_tel_2', 'cus_tax_id'];
-            array_walk($fieldsToClean, function ($field) use (&$data_input) {
-                if (isset($data_input[$field])) {
-                    $data_input[$field] = preg_replace('/[^0-9]/', '', $data_input[$field]);
-                }
-            });
+            // Clean phone numbers & tax ID
+            $this->customer_service->sanitizeNumbers($data_input);
 
             $customer->fill($data_input);
             $customer->cus_id = Str::uuid();
@@ -406,13 +401,8 @@ class CustomerController extends Controller
 
         $data_input = $request->all();
 
-        // Clean เบอร์โทรศัพท์ & Tax ID อัตโนมัติ
-        $fieldsToClean = ['cus_tel_1', 'cus_tel_2', 'cus_tax_id'];
-        array_walk($fieldsToClean, function ($field) use (&$data_input) {
-            if (isset($data_input[$field])) {
-                $data_input[$field] = preg_replace('/[^0-9]/', '', $data_input[$field]);
-            }
-        });
+        // Clean phone numbers & tax ID
+        $this->customer_service->sanitizeNumbers($data_input);
 
 
         try {
