+++ tnp-frontend/src/features/Customer/customerUtils.js
@@ -2,6 +2,7 @@ import moment from "moment";
 import dayjs from "dayjs";
 import "dayjs/locale/th";
 import relativeTime from "dayjs/plugin/relativeTime";
+export { genCustomerNo } from "../../utils/utilityFunction";
 
 // Setup dayjs plugins
 dayjs.extend(relativeTime);
@@ -86,18 +87,3 @@ export const formatRecallDays = (days) => {
   }
 };
 
-export function genCustomerNo(lastCustomerNumber = null) {
-  const currentYear = moment().year().toString();
-
-  let nextId;
-  if (lastCustomerNumber) {
-    const lastYear = lastCustomerNumber.substring(0, 4);
-    const lastId = parseInt(lastCustomerNumber.substring(4), 10);
-
-    nextId = lastYear === currentYear ? lastId + 1 : 1;
-  } else {
-    nextId = 1;
-  }
-
-  return `${currentYear}${nextId.toString().padStart(6, "0")}`;
-}
