+++ tnp-frontend/src/features/UserManagement/customerUtils.js
@@ -1,4 +1,5 @@
 import moment from "moment";
+export { genCustomerNo } from "../../utils/utilityFunction";
 
 // แปลงค่าวันที่ตามลูกค้า ให้อยู่ในรูปแบบนับเวลาถอยหลัง
 export function formatCustomRelativeTime(dateString) {
@@ -13,19 +14,3 @@ export function formatCustomRelativeTime(dateString) {
   }
 }
 
-export function genCustomerNo(lastCustomerNumber = null)
-{
-  const currentYear = moment().year().toString();
-
-  let nextId;
-  if (lastCustomerNumber) {
-    const lastYear = lastCustomerNumber.substring(0, 4);
-    const lastId = parseInt(lastCustomerNumber.substring(4), 10);
-
-    nextId = lastYear === currentYear ? lastId + 1 : 1;
-  } else {
-    nextId = 1;
-  }
-
-  return `${currentYear}${nextId.toString().padStart(6, "0")}`;
-};
